requirements.txt
Flask==3.1.1
pymongo==4.13.2
python-dotenv==1.1.1
Werkzeug==3.1.3
blinker>=1.9.0
PyJWT==2.8.0
waitress==3.0.0